Refactoring of the roles to install Kubernetes and CLIs
