+++ ansible.cfg
@@ -3,4 +3,3 @@ inventory = hosts.yaml
 host_key_checking = False
 remote_user = centos
 roles_path = ./roles
-vault_password_file = .vault
