--- group_vars/all/vars.example
@@ -1,49 +1,0 @@
----
-# Install latest operating system updates
-os_system_updates: False
-
-# DC/OS cluster version
-dcos_version: '1.11.0'
-# Download URL for DC/OS
-dcos_download: "https://downloads.dcos.io/dcos/stable/{{ dcos_version }}/dcos_generate_config.sh"
-
-# Name of the DC/OS Cluster
-dcos_cluster_name: 'demo'
-
-# If planning to upgrade a previous deployed DC/OS Cluster,
-# uncomment the following variable
-#dcos_upgrade_from_version: '1.11.0'
-
-# Deploy Mesosphere Enterprise DC/OS or DC/OS OSS?
-dcos_deploy_ee_package: False
-
-# Optional if dcos_iaas_target := aws
-#dcos_exhibitor: 'aws_s3'
-#dcos_aws_access_key_id: '******'
-#dcos_aws_secret_access_key: '******'
-#dcos_aws_region: 'us-west-2'
-#dcos_s3_bucket: 'bucket-name'
-
-# Optional if dcos_iaas_target := azure
-#dcos_exhibitor: 'azure'
-#dcos_exhibitor_azure_account_name: 'name'
-#dcos_exhibitor_azure_account_key: '******'
-
-# Only required when deploying Mesosphere Enterprise DC/OS
-dcos_ee_security: 'permissive'
-dcos_ee_license_key_contents: '******'
-dcos_ee_superuser_username: admin
-# Default password:= admin
-dcos_ee_superuser_password_hash: "$6$rounds=656000$8CXbMqwuglDt3Yai$ZkLEj8zS.GmPGWt.dhwAv0.XsjYXwVHuS9aHh3DMcfGaz45OpGxC5oQPXUUpFLMkqlXCfhXMloIzE0Xh8VwHJ."
-
-# Configure rexray to enable support of external volumes (only for Mesosphere Enterprise DC/OS)
-# Note: Set rexray_config_method: file and edit ./roles/bootstrap/templates/rexray.yaml.j2 for a custom rexray configuration
-# options: empty, file
-dcos_ee_rexray_config_method: empty
-
-# Install k8s cluster on DC/OS as a service at the end of the DC/OS installation process
-dcos_k8s_enabled: false
-# Kubernetes on DC/OS package version
-dcos_k8s_package_version: '1.0.2-1.9.6'
-# Create SSH tunnel to kubernetes on DC/OS for advanced kubectl CLI commands (like kubectl proxy)
-dcos_k8s_cli_tunnel: true
