+++ plays/install.yml
@@ -1,15 +1,6 @@
 ---
-- name: added localhost to in-memory inventory
-  hosts: all
-  gather_facts: no
-  tasks:
-    - add_host:
-        hostname: localhost
-        ansible_connection: local
-        ansible_host: 127.0.0.1
-
 - name: setup the system requirements on all nodes
-  hosts: all:!localhost
+  hosts: all
   serial: 20
   become: true
   roles:
@@ -29,12 +20,3 @@
   become: true
   roles:
     - node-install
-
-- name: "setup CLIs on control machine{{ ' and install kubernetes as a service on DC/OS' if dcos_k8s_enabled else '' }}"
-  hosts: localhost
-  serial: 1
-  environment:
-    PATH: '..:{{ ansible_env.PATH }}'
-  roles:
-    - { role: cli, when: not dcos_k8s_enabled }
-    - { role: kubernetes, when: dcos_k8s_enabled }
