--- plays/k8s-update.yml
@@ -1,20 +1,0 @@
----
-# This playbook updates kubernetes on DC/OS as a service
-
-- name: added localhost to in-memory inventory
-  hosts: all
-  gather_facts: no
-  tasks:
-    - add_host:
-        hostname: localhost
-        ansible_connection: local
-        ansible_host: 127.0.0.1
-
-- name: update kubernetes as a service on DC/OS
-  hosts: localhost
-  serial: 1
-  environment:
-    PATH: '..:{{ ansible_env.PATH }}'
-
-  roles:
-    - { role: kubernetes, dcos_k8s_enabled: true, dcos_k8s_update: true }
