--- resources/options.json.azure
@@ -1,7 +1,0 @@
-{
-  "kubernetes": {
-    "high_availability": true,
-    "node_count": 3,
-    "public_node_count": 1
-  }
-}
