--- roles/cli/tasks/dcos-uninstall.yml
@@ -1,26 +1,0 @@
----
-# this playbook uninstalls DC/OS CLI
-
-- name: check if DC/OS Enterprise CLI is installed
-  shell: "dcos package list | grep '/dcos-enterprise-cli' | awk '{print $1}'" 
-  register: dcos_ee_cli_installed
-
-- name: uninstall the DC/OS Enterprise CLI
-  expect:
-    command: "dcos package uninstall --app-id=/dcos-enterprise-cli dcos-enterprise-cli"
-    responses:
-      (?i)service: "/dcos-enterprise-cli"
-  when: dcos_ee_cli_installed.stdout == "dcos-enterprise-cli"
-
-- name: delete the hidden .dcos directory
-  become: false
-  file:
-    path: '~/.dcos'
-    state: absent
-  when: cli_uninstall_hidden_config
-
-- name: delete the dcos executable
-  become: false
-  file:
-    path: '../dcos'
-    state: absent
