--- roles/cli/tasks/kubectl-install.yml
@@ -1,16 +1,0 @@
----
-# this playbook installs k8s CLI
-
-- name: lookup latest stable kubernetes version
-  set_fact:
-     kubernetes_version: "{{ lookup('url', 'https://storage.googleapis.com/kubernetes-release/release/stable.txt', split_lines=False) | regex_replace('^(.*)\\n$', '\\1') }}"
-
-- name: download the k8s CLI Linux binary (kubectl) to your current directory and make it executable
-  become: false
-  get_url: url="https://storage.googleapis.com/kubernetes-release/release/{{ kubernetes_version }}/bin/linux/amd64/kubectl" dest="../kubectl" mode=0755 force=yes
-  when: ansible_os_family == "Debian" or ansible_os_family == "RedHat"
-
-- name: download the k8s CLI Darwin binary (kubectl) to your current directory and make it executable
-  become: false
-  get_url: url="https://storage.googleapis.com/kubernetes-release/release/{{ kubernetes_version }}/bin/darwin/amd64/kubectl" dest="../kubectl" mode=0755 force=yes
-  when: ansible_os_family == "Darwin"
