--- roles/cli/tasks/main.yml
@@ -1,22 +1,0 @@
----
-
-- name: connect to cluster
-  dcos_connection:
-    url: "https://{{ groups['masters'][0] }}"
-    insecure: true
-
-- name: create service account for kubernetes
-  dcos_serviceaccount:
-    name: kubernetes
-    description: 'kubernetes service account'
-    secret: kubernetes/secret
-    group: superusers
-    state: present
-
-- name: ensure kubernetes is installed
-  dcos_package:
-    name: kubernetes
-    app_id: kubernetes
-    version: 1.0.2-1.9.6
-    state: present
-    options: { 'service': { 'name': 'kubernetes', "service_account": "kubernetes", "service_account_secret": "kubernetes/secret" }, 'kubernetes': { 'high_availability': true, 'node_count': 3, 'public_node_count': 1 } }
