--- roles/cli/templates/openid-token.j2
@@ -1,2 +1,0 @@
----
-dcos_cli_openid_token: "{{ dcos_cli_openid_token | default('') }}"
