--- roles/kubernetes/dcos-k8s/tasks/upgrade-ee.yml
@@ -1,9 +1,0 @@
----
-# This playbook upgrade kubernetes on Enterprise DC/OS as a service
-
-- name: upgrade kubernetes cluster on Enterprise DC/OS as a service
-  shell: "{{ item }}"
-  with_items:
-    - "dcos package install kubernetes --yes --cli --package-version={{ dcos_k8s_package_version }}"
-    - "dcos kubernetes describe > {{ dcos_k8s_options_file }}"
-    - "dcos kubernetes update --yes --package-version={{ dcos_k8s_package_version }}"
