--- roles/kubernetes/dcos-k8s/templates/options.j2
@@ -1,12 +1,0 @@
-{
-  "kubernetes": {
-    "cloud_provider": "{% if dcos_iaas_target == "aws" %}aws{% else %}(none){% endif %}",
-    "high_availability": true,
-    "node_count": 3,
-    "public_node_count": 1
-  }{% if dcos_deploy_ee_package %},
-  "service": {
-    "service_account": "kubernetes",
-    "service_account_secret": "kubernetes/sa"
-  }{% endif %}
-}
