+++ roles/package/kubernetes/defaults/main.yml
@@ -1,0 +1,8 @@
+---
+dcos_deploy_ee_package: false
+dcos_k8s_enabled: true
+dcos_k8s_cli_tunnel: false
+
+dcos_k8s_app_id: "kubernetes"
+dcos_k8s_proxy_app_id: "{{ dcos_k8s_app_id }}-proxy"
+dcos_k8s_package_version: '1.0.2-1.9.6'
