+++ roles/package/kubernetes/meta/main.yml
@@ -1,0 +1,3 @@
+---
+dependencies:
+  - role: package
