+++ roles/package/tasks/kubectl-install.yml
@@ -1,0 +1,26 @@
+---
+# this playbook installs k8s CLI
+
+- name: create temporay file
+  become: false
+  get_url: url="https://storage.googleapis.com/kubernetes-release/release/stable.txt" dest="../kubectl-version" mode=0744 force=yes
+
+- name: detect kubernetes version
+  shell: cat "../kubectl-version"
+  register: kubernetes_version
+
+- name: download the k8s CLI Linux binary (kubectl) to your current directory and make it executable
+  become: false
+  get_url: url="https://storage.googleapis.com/kubernetes-release/release/{{ kubernetes_version.stdout }}/bin/linux/amd64/kubectl" dest="../kubectl" mode=0755 force=yes
+  when: ansible_os_family == "Debian" or ansible_os_family == "RedHat"
+
+- name: download the k8s CLI Darwin binary (kubectl) to your current directory and make it executable
+  become: false
+  get_url: url="https://storage.googleapis.com/kubernetes-release/release/{{ kubernetes_version.stdout }}/bin/darwin/amd64/kubectl" dest="../kubectl" mode=0755 force=yes
+  when: ansible_os_family == "Darwin"
+
+- name: clean-up temporay file
+  become: false
+  file:
+    path: '../kubectl-version'
+    state: absent
