+++ roles/package/tasks/kubectl-uninstall.yml
@@ -1,0 +1,8 @@
+---
+# this playbook uninstalls k8s CLI
+
+- name: delete the kubectl executable
+  become: false
+  file:
+    path: '../kubectl'
+    state: absent
