--- scripts/get_cli
@@ -1,26 +1,0 @@
-#!/bin/bash
-
-# download dcos and kubectl CLI tools
-
-# check OS version
-unamestr=`uname`
-if [[ "$unamestr" == "Linux" ]]
-then
-  OS="linux"
-else
-  OS="darwin"
-fi
-
-#
-echo "Downloading dcos v$DCOS_VERSION cli for $OS"
-curl -o dcos https://downloads.dcos.io/binaries/cli/$OS/x86-64/dcos-$DCOS_VERSION/dcos
-chmod +x dcos
-echo
-#
-echo "Checking for latest kubectl version..."
-KUBERNETES_VERSION=$(curl -Ss https://storage.googleapis.com/kubernetes-release/release/stable.txt)
-echo "Downloading kubectl $KUBERNETES_VERSION for $OS"
-curl -LO https://storage.googleapis.com/kubernetes-release/release/$KUBERNETES_VERSION/bin/$OS/amd64/kubectl
-chmod +x kubectl
-echo "Download has finished !!!"
-echo
